Refactor: move fetch_an functions to the repondeur
